Add doResponsiveMultiView.js
